requirements-dev.txt
# Development dependencies
# Install with: pip install -r requirements-dev.txt

# Include production requirements
-r requirements.txt

# Testing
coverage==7.3.2
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0

# Code quality
flake8==6.1.0
black==23.11.0
isort==5.12.0

# Security
safety==2.3.5
bandit==1.7.5

# Development tools
django-debug-toolbar==4.2.0
django-extensions==3.2.3
ipython==8.17.2

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.8